+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace set-output with recommended env output ([#234](https://github.com/wabarc/wayback/pull/234))
 - Create deployment instructions for Render ([#236](https://github.com/wabarc/wayback/pull/236))
 - Specify dependencies for the distribution package ([#243](https://github.com/wabarc/wayback/pull/243))
+- Make media downloads are domain-specific ([#247](https://github.com/wabarc/wayback/pull/247))
 
 ## [0.18.1] - 2022-10-30